Handle exec being called before a container is initialized correctly.
Running a command like the following:
```
cid=$(./docker run -tid busybox sh) && \
    echo $cid && ./docker exec -ti $cid echo hello
```
would block indefinitely because we would return success to docker, but fail
because the exec succeeds before the container is ready to run.

Solves the problem by blocking the containerd responses until the VM startup
has returned and moved to process wait (or else returning an error if it fails
at this stage). This resolves deadlocks into docker and allows the above
command to succeed.

Closes #279.