Unblocked: acknowledge non-determinism.
There's no need to be deterministic within the containerd structs, since users
can't be. So, each part of the pod/container/process hierarchy has acquired a
`reaping` channel, which is closed once the logic for whether that object should
be reaped becomes true.

TODO: testing

FURTHER WORK: explicitely firing the reaper at the supervisor level with an
event watch can be removed. Instead, each pod spawns a goroutine which blocks
on the pod's reaping channel, and when closed, does the delete from the
supervisor - and so on down the hierarchy.